Improved audit score tracking (#447)
* initial changes to track score on each individual observation

* added a bunch of UXIssue records for when things pass. Also included scores in each issue

Issues will likely be changed in the future to more appropriately reflect their current use

* added extra scores for more granular scoring to audit score

* Email results link (#446)

* added new email for sending results to user when requested

* Moved location of email sending for audit results

Doesn't currently work, but at least the logic isn't scattered everywhere anymore. Unsure if I should call this progress though.

* updated results email and cleaned up some code

* changed results email to send from my account for the moment

* removed drift before doing crawling or extracting data

* small updates

* contrast audit improved for wcag aa and aaa standards

* Lots of small improvements

* updated shutterbug and altered image algorithm using a composite

* improvements to how we gather links for crawling

* cleanup and refactoring

* set composite image to same as full page screenshot

* improved audit flow and made more robust

* Cleaned up code, and improved flows to prepare for further refactoring

* improvements to flow for better performance

added retry capabilities for better resiliency

* final improvements to audit process

System runs much faster now and audits seem to always finish. Everything is setup and ready to lay in logic for pricing

+++ src/main/java/com/looksee/models/audit/ColorData.java
@@ -17,7 +17,7 @@ public class ColorData extends LookseeObject{
 	@SuppressWarnings("unused")
 	private static Logger log = LoggerFactory.getLogger(ColorData.class);
 
-	private float usage_percent;
+	private double usage_percent;
 	private int red;
 	private int green;
 	private int blue;
@@ -258,11 +258,11 @@ public class ColorData extends LookseeObject{
 		return rgb();
 	}
 
-	public float getUsagePercent() {
+	public double getUsagePercent() {
 		return usage_percent;
 	}
 
-	public void setUsagePercent(float usage_percent) {
+	public void setUsagePercent(double usage_percent) {
 		this.usage_percent = usage_percent;
 	}
